fix: 修复 Go 1.25 下 math.MaxInt64 常量在 fmt.Errorf 中溢出错误
Go 1.25 收紧了 untyped constant 溢出检查，math.MaxInt64 作为
untyped int constant 传入 fmt.Errorf 的 %d 参数时会触发编译错误:
'cannot use math.MaxInt64 (untyped int constant...) as int value
(overflows)'。显式转换为 int64 类型即可解决。

+++ golib/config/config.go
@@ -191,7 +191,7 @@ func parseByteSize(s string) (int64, error) {
 	// 检查整数溢出
 	result := num * float64(multiplier)
 	if result > float64(math.MaxInt64) {
-		return 0, fmt.Errorf("字节大小超出范围: %s (最大支持 %d 字节)", s, math.MaxInt64)
+		return 0, fmt.Errorf("字节大小超出范围: %s (最大支持 %d 字节)", s, int64(math.MaxInt64))
 	}
 
 	return int64(result), nil
